move(user): 파일을 폴더 별로 분리
